Added 'Counter' to Karel
The counter bahaves like a stack.
The counter can be used to introduce the idea of the memory stack or a
stack in general. It can also help students understand what returning a
value from a method means. Also variables can be implemented this way.
It is still open wether the stack should implement convenience
arithmetic operations (incrementValueAtLastSlot, addToValueAtLastSlot:,
...?)

A simple test example (KCCountingKarel) was created.

A new basic structure class (KCFlexibleArray) was added to support the
lazy instanciation of the views representing the individual slots of
the stack. KCFlexibleArray is atm implemented with the usage of a
KCMatrix. (abstracting away that there are no 0 indeces in the matrix)